// src/services/enums/cache.ts
export enum ComputedCacheKey {
  STRING_TO_DATE = 'STRING_TO_DATE',
  DATE_TO_STRING = 'DATE_TO_STRING',
  MONTH_NAME = 'MONTH_NAME',
  DAY_OF_WEEK = 'DAY_OF_WEEK',
  MONTH_FROM_DATE = 'MONTH_FROM_DATE',
  YEAR_FROM_DATE = 'YEAR_FROM_DATE',
  CURRENCY_FORMAT = 'CURRENCY_FORMAT',
  PNL_CLASS = 'PNL_CLASS',
  TRADE_COUNT_FORMAT = 'TRADE_COUNT_FORMAT',
  CALENDAR_DATA = 'CALENDAR_DATA',
  EMPTY_CALENDAR = 'EMPTY_CALENDAR'
}